fix system interface lacking current gem data
